Use unboxed function for building integer based community results

+++ contrib/algo/src/main/java/org/neo4j/graphalgo/TriangleProc.java
@@ -243,7 +243,7 @@ public class TriangleProc {
             return Stream.of(builder.buildLI(graph.nodeCount(), triangles::get));
         } else if (algorithm instanceof TriangleCountQueue){
             final AtomicIntegerArray triangles = ((TriangleCountQueue) algorithm).getTriangles();
-            return Stream.of(builder.buildII(graph.nodeCount(), triangles::get));
+            return Stream.of(builder.buildII(Math.toIntExact(graph.nodeCount()), triangles::get));
         }
         throw new UnsupportedOperationException("unknown algorithm");
     }
@@ -388,7 +388,7 @@ public class TriangleProc {
                 .withTriangleCount(triangleCount.getTriangleCount());
 
         final AtomicIntegerArray triangles = triangleCount.getTriangles();
-        return Stream.of(builder.buildII(graph.nodeCount(), triangles::get));
+        return Stream.of(builder.buildII(Math.toIntExact(graph.nodeCount()), triangles::get));
     }
